[2/2] OmniJaws: set sail to new shores
add broadcast org.omnirom.omnijaws.WEATHER_UPDATE
for clients to listen whenever weather data has
been updated

add new start intent org.omnirom.omnijaws.ACTION_ENABLE
with extra enable true/false to toggle service
enable will always trigger imediate update

Split settings activity with and without icon pack
and enable switch

Remove local tile services

Use correct time to check if update is due based
on alarm time. Forced updates should not influence
the regular alarm schedule

Add omnirom as app icon
