+++ AndroidManifest.xml
@@ -1,72 +1,85 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-	package="org.omnirom.omnijaws"
-	android:versionCode="1"
-	android:versionName="1.0" >
+    package="org.omnirom.omnijaws"
+    android:versionCode="1"
+    android:versionName="1.0" >
 
-	<uses-permission android:name="android.permission.WAKE_LOCK" />
-	<uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
-	<uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-	<uses-permission android:name="android.permission.INTERNET" />
-	<uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
+    <uses-permission android:name="android.permission.WAKE_LOCK" />
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
+    <uses-permission android:name="android.permission.WRITE_SETTINGS" />
 
-	<permission android:name="org.omnirom.omnijaws.READ_WEATHER" />
+    <permission android:name="org.omnirom.omnijaws.READ_WEATHER" />
 
-	<uses-sdk
-		android:minSdkVersion="24"
-		android:targetSdkVersion="25" />
+    <uses-sdk
+        android:minSdkVersion="24"
+        android:targetSdkVersion="25" />
 
-	<application
-		android:allowBackup="true"
-		android:label="@string/app_name" >
+    <application
+        android:allowBackup="true"
+        android:label="@string/app_name"
+        android:icon="@drawable/omnirom_logo" >
 
-		<uses-library android:name="org.apache.http.legacy" android:required="false" />
+        <uses-library android:name="org.apache.http.legacy" android:required="false" />
 
-		<activity
-			android:name=".SettingsActivity"
-			android:exported="true"
-			android:label="@string/settings_title"
-			android:launchMode="singleTop"
-			android:theme="@android:style/Theme.Material.Settings" >
-			<intent-filter>
-				<action android:name="android.intent.action.MAIN" />
-			</intent-filter>
-		</activity>
+        <activity
+            android:name=".SettingsActivity"
+            android:exported="true"
+            android:label="@string/settings_title"
+            android:launchMode="singleTop"
+            android:theme="@android:style/Theme.Material.Settings" >
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+            </intent-filter>
+        </activity>
 
-		<service
-			android:name=".WeatherService"
-			android:exported="true" />
+        <activity
+            android:name=".SettingsActivityService"
+            android:exported="true"
+            android:label="@string/settings_title"
+            android:launchMode="singleTop"
+            android:theme="@android:style/Theme.Material.Settings" >
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+            </intent-filter>
+        </activity>
 
-		<provider
-			android:name=".WeatherContentProvider"
-			android:authorities="org.omnirom.omnijaws.provider"
-			android:exported="true"
-			android:readPermission="org.omnirom.omnijaws.READ_WEATHER" />
+        <service
+            android:name=".WeatherService"
+            android:exported="true" />
 
-		<receiver android:name=".SystemReceiver" >
-			<intent-filter>
-				<action android:name="android.intent.action.BOOT_COMPLETED" />
-			</intent-filter>
-		</receiver>
+        <provider
+            android:name=".WeatherContentProvider"
+            android:authorities="org.omnirom.omnijaws.provider"
+            android:exported="true"
+            android:readPermission="org.omnirom.omnijaws.READ_WEATHER" />
+
+        <receiver android:name=".SystemReceiver" >
+            <intent-filter>
+                <action android:name="android.intent.action.BOOT_COMPLETED" />
+            </intent-filter>
+        </receiver>
 
         <!-- name is used a prefix +"_" for images e.g. weather_<condition_id>
              weather_foo_<condition_id> -->
-		<activity android:name=".weather"
-			  android:label="Default" >
-		    <intent-filter>
-			    <action android:name="org.omnirom.WeatherIconPack" />
-			    <category android:name="android.intent.category.DEFAULT" />
-		    </intent-filter>
-		</activity>
+        <activity android:name=".weather"
+                android:label="Default" >
+            <intent-filter>
+                <action android:name="org.omnirom.WeatherIconPack" />
+                <category android:name="android.intent.category.DEFAULT" />
+            </intent-filter>
+        </activity>
 
-		<service
-			android:name=".client.WeatherTileService"
-			android:icon="@drawable/weather_na"
-			android:label="@string/weather_tile_title"
-			android:permission="android.permission.BIND_QUICK_SETTINGS_TILE">
-			<intent-filter>
-				<action
-					android:name="android.service.quicksettings.action.QS_TILE"/>
-			</intent-filter>
-		</service>
-	 </application>
+        <!--<service
+            android:name=".client.WeatherTileService"
+            android:icon="@drawable/weather_na"
+            android:label="@string/weather_tile_title"
+            android:permission="android.permission.BIND_QUICK_SETTINGS_TILE">
+            <intent-filter>
+                <action
+                    android:name="android.service.quicksettings.action.QS_TILE"/>
+            </intent-filter>
+        </service>-->
+    </application>
 </manifest>
