+++ res/xml/settings.xml
@@ -17,53 +17,47 @@
 -->
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android" >
 
-	<SwitchPreference
-		android:defaultValue="false"
-		android:key="enable"
-		android:title="@string/enable_title" />
+    <SwitchPreference
+        android:defaultValue="false"
+        android:key="enable"
+        android:title="@string/enable_title" />
 
-	<ListPreference
-		android:entries="@array/update_interval_entries"
-		android:entryValues="@array/update_interval_values"
-		android:key="update_interval"
-		android:title="@string/update_interval_title"
-		android:dependency="enable" />
+    <ListPreference
+        android:entries="@array/update_interval_entries"
+        android:entryValues="@array/update_interval_values"
+        android:key="update_interval"
+        android:title="@string/update_interval_title" />
 
-	<Preference
-		android:key="update_status"
-		android:title="@string/last_update_time"
-		android:dependency="enable"
-		android:persistent="false" />
+    <Preference
+        android:key="update_status"
+        android:title="@string/last_update_time"
+        android:persistent="false" />
 
-	<ListPreference
-		android:entries="@array/provider_entries"
-		android:entryValues="@array/provider_values"
-		android:key="provider"
-		android:title="@string/provider_title"
-		android:dependency="enable" />
+    <ListPreference
+        android:entries="@array/provider_entries"
+        android:entryValues="@array/provider_values"
+        android:key="provider"
+        android:title="@string/provider_title" />
 
-	<ListPreference
-		android:entries="@array/units_entries"
-		android:entryValues="@array/units_values"
-		android:key="units"
-		android:title="@string/units_title"
-		android:dependency="enable" />
+    <ListPreference
+        android:entries="@array/units_entries"
+        android:entryValues="@array/units_values"
+        android:key="units"
+        android:title="@string/units_title" />
 
-	<CheckBoxPreference
-		android:defaultValue="false"
-		android:key="custom_location"
-		android:title="@string/custom_location_title"
-		android:dependency="enable" />
+    <CheckBoxPreference
+        android:defaultValue="false"
+        android:key="custom_location"
+        android:title="@string/custom_location_title" />
 
-	<org.omnirom.omnijaws.CustomLocationPreference
-		android:dependency="custom_location"
-		android:dialogTitle="@string/weather_custom_location_dialog_title"
-		android:key="weather_custom_location_city"
-		android:title="@string/weather_custom_location_title" />
+    <org.omnirom.omnijaws.CustomLocationPreference
+        android:dependency="custom_location"
+        android:dialogTitle="@string/weather_custom_location_dialog_title"
+        android:key="weather_custom_location_city"
+        android:title="@string/weather_custom_location_title" />
 
-	<ListPreference
-		android:key="weather_icon_pack"
-		android:title="@string/weather_icon_pack_title"
-		android:dependency="enable" />
+    <ListPreference
+        android:key="weather_icon_pack"
+        android:title="@string/weather_icon_pack_title" />
 
 </PreferenceScreen>
