+++ src/org/omnirom/omnijaws/SettingsActivity.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2015 The OmniROM Project
+ *  Copyright (C) 2017 The OmniROM Project
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,368 +17,13 @@
  */
 package org.omnirom.omnijaws;
 
-import android.Manifest;
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.content.Context;
-import android.content.DialogInterface;
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.content.pm.PackageManager;
-import android.content.pm.ResolveInfo;
-import android.location.LocationManager;
-import android.os.AsyncTask;
 import android.os.Bundle;
-import android.os.Handler;
-import android.preference.CheckBoxPreference;
-import android.preference.ListPreference;
-import android.preference.Preference;
-import android.preference.Preference.OnPreferenceChangeListener;
-import android.preference.PreferenceActivity;
-import android.preference.PreferenceManager;
-import android.preference.PreferenceScreen;
-import android.preference.SwitchPreference;
-import android.provider.Settings;
-import android.util.Log;
-import android.view.MenuItem;
 
-import org.omnirom.omnijaws.client.OmniJawsClient;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class SettingsActivity extends PreferenceActivity implements OnPreferenceChangeListener, WeatherLocationTask.Callback  {
-
-    private static final String WEATHER_SERVICE_PACKAGE = "org.omnirom.omnijaws";
-    private static final String CHRONUS_ICON_PACK_INTENT = "com.dvtonder.chronus.ICON_PACK";
-    private static final String DEFAULT_WEATHER_ICON_PACKAGE = "org.omnirom.omnijaws";
-
-    private SharedPreferences mPrefs;
-    private ListPreference mProvider;
-    private CheckBoxPreference mCustomLocation;
-    private ListPreference mUnits;
-    private SwitchPreference mEnable;
-    private boolean mTriggerUpdate;
-    private boolean mTriggerPermissionCheck;
-    private ListPreference mUpdateInterval;
-    private CustomLocationPreference mLocation;
-    private ListPreference mWeatherIconPack;
-    private Preference mUpdateStatus;
-    private Handler mHandler = new Handler();
-
-    private static final String PREF_KEY_CUSTOM_LOCATION_CITY = "weather_custom_location_city";
-    private static final int PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION = 0;
-    private static final String WEATHER_ICON_PACK = "weather_icon_pack";
-    private static final String PREF_KEY_UPDATE_STATUS = "update_status";
+public class SettingsActivity extends SettingsActivityService  {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
+        mShowIconPack = true;
         super.onCreate(savedInstanceState);
-        mPrefs = PreferenceManager.getDefaultSharedPreferences(this);
-        getActionBar().setDisplayHomeAsUpEnabled(true);
-
-        addPreferencesFromResource(R.xml.settings);
-
-        mEnable = (SwitchPreference) findPreference(Config.PREF_KEY_ENABLE);
-        mCustomLocation = (CheckBoxPreference) findPreference(Config.PREF_KEY_CUSTOM_LOCATION);
-
-        mProvider = (ListPreference) findPreference(Config.PREF_KEY_PROVIDER);
-        mProvider.setOnPreferenceChangeListener(this);
-        int idx = mProvider.findIndexOfValue(mPrefs.getString(Config.PREF_KEY_PROVIDER,
-                mProvider.getEntryValues()[0].toString()));
-        mProvider.setValueIndex(idx);
-        mProvider.setSummary(mProvider.getEntries()[idx]);
-
-        mUnits = (ListPreference) findPreference(Config.PREF_KEY_UNITS);
-        mUnits.setOnPreferenceChangeListener(this);
-        idx = mUnits.findIndexOfValue(mPrefs.getString(Config.PREF_KEY_UNITS,
-                mUnits.getEntryValues()[0].toString()));
-        mUnits.setValueIndex(idx);
-        mUnits.setSummary(mUnits.getEntries()[idx]);
-
-        mUpdateInterval = (ListPreference) findPreference(Config.PREF_KEY_UPDATE_INTERVAL);
-        mUpdateInterval.setOnPreferenceChangeListener(this);
-        idx = mUpdateInterval.findIndexOfValue(mPrefs.getString(Config.PREF_KEY_UPDATE_INTERVAL,
-                mUpdateInterval.getEntryValues()[0].toString()));
-        mUpdateInterval.setValueIndex(idx);
-        mUpdateInterval.setSummary(mUpdateInterval.getEntries()[idx]);
-
-        mLocation = (CustomLocationPreference) findPreference(PREF_KEY_CUSTOM_LOCATION_CITY);
-        if (mPrefs.getBoolean(Config.PREF_KEY_ENABLE, false)
-                && !mPrefs.getBoolean(Config.PREF_KEY_CUSTOM_LOCATION, false)) {
-            mTriggerUpdate = false;
-            checkLocationEnabled();
-        }
-        mWeatherIconPack = (ListPreference) findPreference(WEATHER_ICON_PACK);
-
-        String settingHeaderPackage = Config.getIconPack(this);
-        List<String> entries = new ArrayList<String>();
-        List<String> values = new ArrayList<String>();
-        getAvailableWeatherIconPacks(entries, values);
-        mWeatherIconPack.setEntries(entries.toArray(new String[entries.size()]));
-        mWeatherIconPack.setEntryValues(values.toArray(new String[values.size()]));
-
-        int valueIndex = mWeatherIconPack.findIndexOfValue(settingHeaderPackage);
-        if (valueIndex == -1) {
-            // no longer found
-            settingHeaderPackage = DEFAULT_WEATHER_ICON_PACKAGE;
-            Config.setIconPack(this, settingHeaderPackage);
-            valueIndex = mWeatherIconPack.findIndexOfValue(settingHeaderPackage);
-        }
-        mWeatherIconPack.setValueIndex(valueIndex >= 0 ? valueIndex : 0);
-        mWeatherIconPack.setSummary(mWeatherIconPack.getEntry());
-        mWeatherIconPack.setOnPreferenceChangeListener(this);
-
-        mUpdateStatus = findPreference(PREF_KEY_UPDATE_STATUS);
-        queryLastUpdateTime();
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        if (mTriggerPermissionCheck) {
-            checkLocationPermissions();
-            mTriggerPermissionCheck = false;
-        }
-    }
-
-    @Override
-    public boolean onPreferenceTreeClick(PreferenceScreen preferenceScreen,
-            Preference preference) {
-        if (preference == mCustomLocation) {
-            if (!mCustomLocation.isChecked()) {
-                mTriggerUpdate = true;
-                checkLocationEnabled();
-            } else {
-                if (Config.getLocationName(this) != null) {
-                    // city ids are provider specific - so we need to recheck
-                    // cause provider migth be changed while unchecked
-                    new WeatherLocationTask(this, Config.getLocationName(this), this).execute();
-                } else {
-                    disableService();
-                }
-            }
-            return true;
-        } else if (preference == mEnable) {
-            if (mEnable.isChecked()) {
-                if (!mCustomLocation.isChecked()) {
-                    mTriggerUpdate = true;
-                    checkLocationEnabled();
-                } else {
-                    WeatherService.scheduleUpdate(this);
-                }
-            } else {
-                disableService();
-            }
-            queryLastUpdateTime();
-            return true;
-        } else if (preference == mUpdateStatus) {
-            WeatherService.startUpdate(this, true);
-            queryLastUpdateTime();
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public boolean onPreferenceChange(Preference preference, Object newValue) {
-        if (preference == mProvider) {
-            String value = (String) newValue;
-            int idx = mProvider.findIndexOfValue(value);
-            mProvider.setSummary(mProvider.getEntries()[idx]);
-            mProvider.setValueIndex(idx);
-            if (mCustomLocation.isChecked() && Config.getLocationName(this) != null) {
-                // city ids are provider specific - so we need to recheck
-                new WeatherLocationTask(this, Config.getLocationName(this), this).execute();
-            } else {
-                WeatherService.startUpdate(this, true);
-            }
-            return true;
-        } else if (preference == mUnits) {
-            String value = (String) newValue;
-            int idx = mUnits.findIndexOfValue(value);
-            mUnits.setSummary(mUnits.getEntries()[idx]);
-            mUnits.setValueIndex(idx);
-            WeatherService.startUpdate(this, true);
-            return true;
-        } else if (preference == mUpdateInterval) {
-            String value = (String) newValue;
-            int idx = mUpdateInterval.findIndexOfValue(value);
-            mUpdateInterval.setSummary(mUpdateInterval.getEntries()[idx]);
-            mUpdateInterval.setValueIndex(idx);
-            WeatherService.scheduleUpdate(this);
-            queryLastUpdateTime();
-            return true;
-        } else if (preference == mWeatherIconPack) {
-            String value = (String) newValue;
-            Config.setIconPack(this, value);
-            int valueIndex = mWeatherIconPack.findIndexOfValue(value);
-            mWeatherIconPack.setSummary(mWeatherIconPack.getEntries()[valueIndex]);
-            return true;
-        }
-        return false;
-    }
-    
-    private void showDialog() {
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        final Dialog dialog;
-
-        // Build and show the dialog
-        builder.setTitle(R.string.weather_retrieve_location_dialog_title);
-        builder.setMessage(R.string.weather_retrieve_location_dialog_message);
-        builder.setCancelable(false);
-        builder.setPositiveButton(R.string.weather_retrieve_location_dialog_enable_button,
-                new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int whichButton) {
-                        mTriggerPermissionCheck = true;
-                        mTriggerUpdate = true;
-                        Intent intent = new Intent(Settings.ACTION_LOCATION_SOURCE_SETTINGS);
-                        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                        startActivity(intent);
-                    }
-                });
-        builder.setNegativeButton(android.R.string.cancel, null);
-        dialog = builder.create();
-        dialog.show();
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-        case android.R.id.home:
-            finish();
-            return true;
-        default:
-            break;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
-    private void checkLocationPermissions() {
-        if (checkSelfPermission(Manifest.permission.ACCESS_FINE_LOCATION)
-                != PackageManager.PERMISSION_GRANTED) {
-            requestPermissions(new String[] { Manifest.permission.ACCESS_FINE_LOCATION },
-                    PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION);
-        } else {
-            if (mTriggerUpdate) {
-                WeatherService.scheduleUpdate(this);
-                mTriggerUpdate = false;
-            }
-        }
-    }
-
-    private void checkLocationEnabled() {
-        LocationManager lm = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
-        if (!lm.isProviderEnabled(LocationManager.NETWORK_PROVIDER)) {
-            showDialog();
-        } else {
-            checkLocationPermissions();
-        }
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
-        switch (requestCode) {
-            case PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION: {
-                // If request is cancelled, the result arrays are empty.
-                if (grantResults.length > 0
-                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    if (mTriggerUpdate) {
-                        WeatherService.scheduleUpdate(this);
-                        mTriggerUpdate = false;
-                    }
-                }
-                break;
-            }
-        }
-    }
-
-    private void disableService() {
-        // stop any pending
-        WeatherService.cancelUpdate(this);
-    }
-
-    @Override
-    public void applyLocation(WeatherInfo.WeatherLocation result) {
-        Config.setLocationId(this, result.id);
-        Config.setLocationName(this, result.city);
-        mLocation.setText(result.city);
-        mLocation.setSummary(result.city);
-        WeatherService.startUpdate(this, true);
-    }
-
-    private void getAvailableWeatherIconPacks(List<String> entries, List<String> values) {
-        Intent i = new Intent();
-        PackageManager packageManager = getPackageManager();
-        i.setAction("org.omnirom.WeatherIconPack");
-        for (ResolveInfo r : packageManager.queryIntentActivities(i, 0)) {
-            String packageName = r.activityInfo.packageName;
-            if (packageName.equals(DEFAULT_WEATHER_ICON_PACKAGE)) {
-                values.add(0, r.activityInfo.name);
-            } else {
-                values.add(r.activityInfo.name);
-            }
-            String label = r.activityInfo.loadLabel(getPackageManager()).toString();
-            if (label == null) {
-                label = r.activityInfo.packageName;
-            }
-            if (packageName.equals(DEFAULT_WEATHER_ICON_PACKAGE)) {
-                entries.add(0, label);
-            } else {
-                entries.add(label);
-            }
-        }
-        i = new Intent(Intent.ACTION_MAIN);
-        i.addCategory(CHRONUS_ICON_PACK_INTENT);
-        for (ResolveInfo r : packageManager.queryIntentActivities(i, 0)) {
-            String packageName = r.activityInfo.packageName;
-            values.add(packageName + ".weather");
-            String label = r.activityInfo.loadLabel(getPackageManager()).toString();
-            if (label == null) {
-                label = r.activityInfo.packageName;
-            }
-            entries.add(label);
-        }
-    }
-
-    private void queryLastUpdateTime() {
-        final AsyncTask<Void, Void, Void> t = new AsyncTask<Void, Void, Void>() {
-            @Override
-            protected void onProgressUpdate(Void... values) {
-            }
-            @Override
-            protected Void doInBackground(Void... params) {
-                final String updateTime = getLastUpdateTime();
-                SettingsActivity.this.runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        mUpdateStatus.setSummary(updateTime);
-                    }
-                });
-                return null;
-            }
-        };
-        mHandler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                t.execute();
-            }
-        }, 2000);
-    }
-
-    private String getLastUpdateTime() {
-        OmniJawsClient mWeatherClient = new OmniJawsClient(this);
-        if (mWeatherClient.isOmniJawsEnabled()) {
-            OmniJawsClient.WeatherInfo mWeatherData = null;
-            try {
-                mWeatherClient.queryWeather();
-                mWeatherData = mWeatherClient.getWeatherInfo();
-                if (mWeatherData != null) {
-                    return mWeatherData.getLastUpdateTime();
-                }
-            } catch(Exception e) {
-            }
-        }
-        return getResources().getString(R.string.service_disabled);
     }
 }
